Refactor LaunchManager for no EAC via Steam on Linux
