+++ RLBotCSTests/TestTomls/default.toml
@@ -86,7 +86,7 @@ demolish_score = "Zero"
 # "One", "Zero", "Two", "Three", "Five", "Ten"
 normal_goal_score = "One"
 # "One", "Zero", "Two", "Three", "Five", "Ten"
-aerial_goal_score =  "One"
+aerial_goal_score = "One"
 # "Zero", "One", "Two", "Three"
 assist_goal_score = "Zero"
 # "Default", "Backwards"
